MVP/backend/django_alternative/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
python-dotenv>=1.0,<2.0
anthropic>=0.25,<1.0
openpyxl>=3.1,<4.0
psycopg2-binary>=2.9,<3.0
